pkp/pkp-lib#12841 refactor exception handler to own class
